+++ client/app/modules/settings/authentication/authentication.component.css
@@ -1,0 +1,4 @@
+.ui.logo.image {
+  width: 50px !important; }
+
+/*# sourceMappingURL=authentication.component.css.map */
