+++ client/app/modules/settings/authentication/authentication.component.scss
@@ -1,0 +1,3 @@
+.ui.logo.image {
+    width: 50px !important;
+}
