+++ client/app/modules/settings/settings.component.css
@@ -1,4 +1,3 @@
-.ui.logo.image {
-  width: 50px !important; }
+
 
 /*# sourceMappingURL=settings.component.css.map */
