+++ client/app/modules/settings/settings.component.html
@@ -2,52 +2,4 @@
     Settings
 </h1>
 
-<div>
-
-</div>
-
-<table class="ui very basic table">
-    <thead>
-    </thead>
-    <tbody>
-    <tr>
-        <td><img class="ui logo image" src="assets/logos/box_logo.png"></td>
-        <td>Online file sharing and personal cloud content management service for businesses</td>
-        <td class="right aligned">
-            <a class="ui button"
-               *ngIf="!user.boxAuthenticated"
-               [class.loading]="authenticatingBox"
-               (click)="navigateBoxAuth()">Authenticate Box</a>
-            <a class="ui positive button"
-               *ngIf="user.boxAuthenticated"
-               (click)="navigateBoxAuth()">Authenticated</a>
-        </td>
-    </tr>
-    <tr>
-        <td><img class="ui logo image" src="assets/logos/trello_logo.png"></td>
-        <td>Organize and prioritize your projects in a fun, flexible and rewarding way</td>
-        <td class="right aligned">
-            <a class="ui button"
-               *ngIf="!user.trelloAuthenticated"
-               [class.loading]="authenticatingTrello"
-               (click)="navigateTrelloAuth()">Authenticate Trello</a>
-            <a class="ui positive button"
-               *ngIf="user.trelloAuthenticated"
-               (click)="navigateTrelloAuth()">Authenticated</a>
-        </td>
-    </tr>
-    <tr>
-        <td><img class="ui logo image" src="assets/logos/slack_logo.png"></td>
-        <td>All the tools your team needs in one place. Messaging, search, archiving & more</td>
-        <td class="right aligned">
-            <a class="ui button"
-               *ngIf="!user.slackAuthenticated"
-               [class.loading]="authenticatingSlack"
-               (click)="navigateSlackAuth()">Authenticate Slack</a>
-            <a class="ui positive button"
-               *ngIf="user.slackAuthenticated"
-               (click)="navigateSlackAuth()">Authenticated</a>
-        </td>
-    </tr>
-    </tbody>
-</table>
+<app-authentication></app-authentication>
