--- client/app/modules/settings/settings.component.scss
@@ -1,3 +1,0 @@
-.ui.logo.image {
-    width: 50px !important;
-}
