+++ client/app/modules/settings/settings.module.ts
@@ -4,13 +4,15 @@ import {FormsModule} from "@angular/forms";
 import {SharedModule} from "../shared/shared.module";
 import {SettingsComponent} from "./settings.component";
 import {routing} from "./settings.routing";
+import {AuthenticationModule} from "./authentication/authentication.module";
 
 @NgModule({
     imports: [
         CommonModule,
         FormsModule,
         SharedModule.forRoot(),
-        routing
+        routing,
+        AuthenticationModule
     ],
     declarations: [SettingsComponent]
 })
